@@ -21,13 +21,16 @@ class NeumorphicColors {
   static const Color defaultBorder = Color(0x33000000);
   static const Color darkDefaultBorder = Color(0x33FFFFFF);
 
-  static const Color decorationMaxWhiteColor =
-      Color(0xFFFFFFFF); //for intensity = 1
-  static const Color decorationMaxDarkColor =
-      Color(0x8A000000); //for intensity = 1
+  static const Color decorationMaxWhiteColor = Color(
+    0xFFFFFFFF,
+  ); //for intensity = 1
+  static const Color decorationMaxDarkColor = Color(
+    0x8A000000,
+  ); //for intensity = 1
 
-  static const Color embossMaxWhiteColor =
-      Color(0x99FFFFFF); //for intensity = 1
+  static const Color embossMaxWhiteColor = Color(
+    0x99FFFFFF,
+  ); //for intensity = 1
   static const Color embossMaxDarkColor = Color(0x73000000); //for intensity = 1
 
   static const Color _gradientShaderDarkColor = Color(0x8A000000);
@@ -35,61 +38,54 @@ class NeumorphicColors {
 
   static const Color defaultTextColor = Color(0xFF000000);
 
-  NeumorphicColors._();
+  const NeumorphicColors._();
 
   static Color decorationWhiteColor(Color color, {required double intensity}) {
     // intensity act on opacity;
-    return _applyPercentageOnOpacity(
-      maxColor: color,
-      percent: intensity,
-    );
+    return _applyPercentageOnOpacity(maxColor: color, percent: intensity);
   }
 
   static Color decorationDarkColor(Color color, {required double intensity}) {
     // intensity act on opacity;
-    return _applyPercentageOnOpacity(
-      maxColor: color,
-      percent: intensity,
-    );
+    return _applyPercentageOnOpacity(maxColor: color, percent: intensity);
   }
 
   static Color embossWhiteColor(Color color, {required double intensity}) {
     // intensity act on opacity;
-    return _applyPercentageOnOpacity(
-      maxColor: color,
-      percent: intensity,
-    );
+    return _applyPercentageOnOpacity(maxColor: color, percent: intensity);
   }
 
   static Color embossDarkColor(Color color, {required double intensity}) {
     // intensity act on opacity;
-    return _applyPercentageOnOpacity(
-      maxColor: color,
-      percent: intensity,
-    );
+    return _applyPercentageOnOpacity(maxColor: color, percent: intensity);
   }
 
   static Color gradientShaderDarkColor({required double intensity}) {
     // intensity act on opacity;
     return _applyPercentageOnOpacity(
-        maxColor: NeumorphicColors._gradientShaderDarkColor,
-        percent: intensity);
+      maxColor: NeumorphicColors._gradientShaderDarkColor,
+      percent: intensity,
+    );
   }
 
   static Color gradientShaderWhiteColor({required double intensity}) {
     // intensity act on opacity;
     return _applyPercentageOnOpacity(
-        maxColor: NeumorphicColors._gradientShaderWhiteColor,
-        percent: intensity);
+      maxColor: NeumorphicColors._gradientShaderWhiteColor,
+      percent: intensity,
+    );
   }
 
-  static Color _applyPercentageOnOpacity(
-      {required Color maxColor, required double percent}) {
-    final maxOpacity = maxColor.opacity;
-    final maxIntensity = Neumorphic.MAX_INTENSITY;
+  static Color _applyPercentageOnOpacity({
+    required Color maxColor,
+    required double percent,
+  }) {
+    final maxOpacity = maxColor.a;
+    const maxIntensity = Neumorphic.MAX_INTENSITY;
     final newOpacity = percent * maxOpacity / maxIntensity;
-    final newColor =
-        maxColor.withOpacity(newOpacity); //<-- intensity act on opacity;
+    final newColor = maxColor.withValues(
+      alpha: newOpacity,
+    ); //<-- intensity act on opacity;
     return newColor;
   }
 }
