@@ -1,33 +1,27 @@
-import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
-
-import 'theme.dart';
-import 'theme_wrapper.dart';
 
 export 'theme.dart';
 export 'theme_wrapper.dart';
 
-typedef NeumorphicThemeUpdater = NeumorphicThemeData Function(
-    NeumorphicThemeData? current);
+typedef NeumorphicThemeUpdater =
+    NeumorphicThemeData Function(NeumorphicThemeData? current);
 
 class NeumorphicThemeInherited extends InheritedWidget {
-  final Widget child;
   final ThemeWrapper value;
   final ValueChanged<ThemeWrapper> onChanged;
 
-  NeumorphicThemeInherited(
-      {Key? key,
-      required this.child,
-      required this.value,
-      required this.onChanged})
-      : super(key: key, child: child);
+  const NeumorphicThemeInherited({
+    super.key,
+    required super.child,
+    required this.value,
+    required this.onChanged,
+  });
 
   @override
   bool updateShouldNotify(NeumorphicThemeInherited old) => value != old.value;
 
   NeumorphicThemeData? get current {
-    return this.value.current;
+    return value.current;
   }
 
   bool get isUsingDark {
@@ -37,18 +31,18 @@ class NeumorphicThemeInherited extends InheritedWidget {
   ThemeMode get themeMode => value.themeMode;
 
   set themeMode(ThemeMode currentTheme) {
-    this.onChanged(value.copyWith(currentTheme: currentTheme));
+    onChanged(value.copyWith(currentTheme: currentTheme));
   }
 
   void updateCurrentTheme(NeumorphicThemeData update) {
     if (value.useDark) {
       final newValue = value.copyWith(darkTheme: update);
       //this.value = newValue;
-      this.onChanged(newValue);
+      onChanged(newValue);
     } else {
       final newValue = value.copyWith(theme: update);
       //this.value = newValue;
-      this.onChanged(newValue);
+      onChanged(newValue);
     }
   }
 
@@ -57,11 +51,11 @@ class NeumorphicThemeInherited extends InheritedWidget {
     if (value.useDark) {
       final newValue = value.copyWith(darkTheme: update);
       //this.value = newValue;
-      this.onChanged(newValue);
+      onChanged(newValue);
     } else {
       final newValue = value.copyWith(theme: update);
       //this.value = newValue;
-      this.onChanged(newValue);
+      onChanged(newValue);
     }
   }
 }
