@@ -2,9 +2,9 @@ import 'dart:ui';
 
 import 'package:flutter/widgets.dart';
 
-import '../neumorphic_box_shape.dart';
-import 'neumorphic_decoration_painter.dart';
-import 'neumorphic_emboss_decoration_painter.dart';
+import 'package:flutter_neumorphic/src/neumorphic_box_shape.dart';
+import 'package:flutter_neumorphic/src/decoration/neumorphic_decoration_painter.dart';
+import 'package:flutter_neumorphic/src/decoration/neumorphic_emboss_decoration_painter.dart';
 
 @immutable
 class NeumorphicDecoration extends Decoration {
@@ -14,7 +14,7 @@ class NeumorphicDecoration extends Decoration {
   final bool renderingByPath;
   final bool isForeground;
 
-  NeumorphicDecoration({
+  const NeumorphicDecoration({
     required this.style,
     required this.isForeground,
     required this.renderingByPath,
@@ -34,7 +34,7 @@ class NeumorphicDecoration extends Decoration {
         //only box draw background
         drawShadow: !isForeground,
         //only box draw shadow
-        renderingByPath: this.renderingByPath,
+        renderingByPath: renderingByPath,
         onChanged: onChanged ?? () {},
         shape: shape,
       );
@@ -53,25 +53,27 @@ class NeumorphicDecoration extends Decoration {
   @override
   NeumorphicDecoration lerpFrom(Decoration? a, double t) {
     if (a == null) return scale(t);
-    if (a is NeumorphicDecoration)
+    if (a is NeumorphicDecoration) {
       return NeumorphicDecoration.lerp(a, this, t)!;
-    return super.lerpFrom(a, t) as NeumorphicDecoration;
+    }
+    return super.lerpFrom(a, t)! as NeumorphicDecoration;
   }
 
   @override
   NeumorphicDecoration lerpTo(Decoration? b, double t) {
     if (b == null) return scale(1.0 - t);
-    if (b is NeumorphicDecoration)
+    if (b is NeumorphicDecoration) {
       return NeumorphicDecoration.lerp(this, b, t)!;
-    return super.lerpTo(b, t) as NeumorphicDecoration;
+    }
+    return super.lerpTo(b, t)! as NeumorphicDecoration;
   }
 
   NeumorphicDecoration scale(double factor) {
-    print("scale");
+    debugPrint("scale");
     return NeumorphicDecoration(
-        isForeground: this.isForeground,
-        renderingByPath: this.renderingByPath,
-        splitBackgroundForeground: this.splitBackgroundForeground,
+        isForeground: isForeground,
+        renderingByPath: renderingByPath,
+        splitBackgroundForeground: splitBackgroundForeground,
         shape: NeumorphicBoxShape.lerp(null, shape, factor)!,
         style: style.copyWith());
   }
@@ -92,8 +94,8 @@ class NeumorphicDecoration extends Decoration {
       return b;
     }
 
-    var aStyle = a.style;
-    var bStyle = b.style;
+    final NeumorphicStyle aStyle = a.style;
+    final NeumorphicStyle bStyle = b.style;
 
     return NeumorphicDecoration(
         isForeground: a.isForeground,
