@@ -1,17 +1,14 @@
-import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart';
 
-import '../neumorphic_box_shape.dart';
-import '../theme/theme.dart';
-import 'cache/neumorphic_emboss_painter_cache.dart';
+import 'package:flutter_neumorphic/src/neumorphic_box_shape.dart';
+import 'package:flutter_neumorphic/src/theme/theme.dart';
+import 'package:flutter_neumorphic/src/decoration/cache/neumorphic_emboss_painter_cache.dart';
 
 export '../theme/theme.dart';
 
 class NeumorphicEmbossDecorationPainter extends BoxPainter {
-  NeumorphicEmbossPainterCache _cache;
+  final NeumorphicEmbossPainterCache _cache;
 
   final NeumorphicStyle style;
   final NeumorphicBoxShape shape;
@@ -32,20 +29,20 @@ class NeumorphicEmbossDecorationPainter extends BoxPainter {
       required this.drawShadow,
       required VoidCallback onChanged,
       NeumorphicBoxShape? shape})
-      : this.shape = shape ?? NeumorphicBoxShape.rect(),
+      : shape = shape ?? const NeumorphicBoxShape.rect(),
         _cache = NeumorphicEmbossPainterCache(),
         super(onChanged) {
     _generatePainters();
   }
 
   void _generatePainters() {
-    this._backgroundPaint = Paint();
-    this._whiteShadowPaint = Paint();
-    this._whiteShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
-    this._blackShadowPaint = Paint();
-    this._blackShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
+    _backgroundPaint = Paint();
+    _whiteShadowPaint = Paint();
+    _whiteShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
+    _blackShadowPaint = Paint();
+    _blackShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
 
-    this._borderPaint = Paint()
+    _borderPaint = Paint()
       ..strokeCap = StrokeCap.round
       ..strokeJoin = StrokeJoin.bevel
       ..style = PaintingStyle.stroke;
@@ -57,8 +54,7 @@ class NeumorphicEmbossDecorationPainter extends BoxPainter {
       required NeumorphicStyle newStyle}) {
     bool invalidateSize = false;
     if (configuration.size != null) {
-      invalidateSize = this
-          ._cache
+      invalidateSize = _cache
           .updateSize(newOffset: offset, newSize: configuration.size!);
       if (invalidateSize) {
         _cache.updatePath(
@@ -68,39 +64,38 @@ class NeumorphicEmbossDecorationPainter extends BoxPainter {
     }
 
     bool invalidateLightSource = false;
-    invalidateLightSource = this
-        ._cache
+    invalidateLightSource = _cache
         .updateLightSource(style.lightSource, style.oppositeShadowLightSource);
 
     bool invalidateColor = false;
     if (style.color != null) {
-      invalidateColor = this._cache.updateStyleColor(style.color!);
+      invalidateColor = _cache.updateStyleColor(style.color!);
       if (invalidateColor) {
-        _backgroundPaint..color = _cache.backgroundColor;
+        _backgroundPaint.color = _cache.backgroundColor;
       }
     }
     bool invalidateDepth = false;
     if (style.depth != null) {
-      invalidateDepth = this._cache.updateStyleDepth(style.depth!, 5);
+      invalidateDepth = _cache.updateStyleDepth(style.depth!, 5);
       if (invalidateDepth) {
-        _blackShadowMaskPaint..maskFilter = _cache.maskFilterBlur;
-        _whiteShadowMaskPaint..maskFilter = _cache.maskFilterBlur;
+        _blackShadowMaskPaint.maskFilter = _cache.maskFilterBlur;
+        _whiteShadowMaskPaint.maskFilter = _cache.maskFilterBlur;
       }
     }
 
-    final bool invalidateShadowColors = this._cache.updateShadowColor(
+    final bool invalidateShadowColors = _cache.updateShadowColor(
           newShadowLightColorEmboss:
-              style.shadowLightColorEmboss ?? Color(0xFFFFFFFF),
+              style.shadowLightColorEmboss ?? const Color(0xFFFFFFFF),
           newShadowDarkColorEmboss:
-              style.shadowDarkColorEmboss ?? Color(0xFF000000),
+              style.shadowDarkColorEmboss ?? const Color(0xFF000000),
           newIntensity: style.intensity ?? 0.25,
         );
     if (invalidateShadowColors) {
       if (_cache.shadowLightColor != null) {
-        _whiteShadowPaint..color = _cache.shadowLightColor!;
+        _whiteShadowPaint.color = _cache.shadowLightColor!;
       }
       if (_cache.shadowDarkColor != null) {
-        _blackShadowPaint..color = _cache.shadowDarkColor!;
+        _blackShadowPaint.color = _cache.shadowDarkColor!;
       }
     }
 
@@ -126,7 +121,7 @@ class NeumorphicEmbossDecorationPainter extends BoxPainter {
         ..drawPath(
             path,
             _borderPaint
-              ..color = style.border.color ?? Color(0x00000000)
+              ..color = style.border.color ?? const Color(0x00000000)
               ..strokeWidth = style.border.width ?? 0)
         ..restore();
     }
@@ -158,8 +153,8 @@ class NeumorphicEmbossDecorationPainter extends BoxPainter {
   @override
   void paint(Canvas canvas, Offset offset, ImageConfiguration configuration) {
     _updateCache(
-        offset: offset, configuration: configuration, newStyle: this.style);
-    for (var subPath in _cache.subPaths) {
+        offset: offset, configuration: configuration, newStyle: style);
+    for (final Path subPath in _cache.subPaths) {
       if (drawBackground) {
         _paintBackground(canvas, subPath);
       }
