@@ -1,13 +1,12 @@
 import 'dart:ui' as ui;
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 
-import '../theme/theme.dart';
-import 'cache/neumorphic_painter_cache.dart';
-import 'neumorphic_box_decoration_helper.dart';
-import 'neumorphic_emboss_decoration_painter.dart';
+import 'package:flutter_neumorphic/src/theme/theme.dart';
+import 'package:flutter_neumorphic/src/decoration/cache/neumorphic_painter_cache.dart';
+import 'package:flutter_neumorphic/src/decoration/neumorphic_box_decoration_helper.dart';
+import 'package:flutter_neumorphic/src/decoration/neumorphic_emboss_decoration_painter.dart';
 
 class NeumorphicEmptyTextPainter extends BoxPainter {
   NeumorphicEmptyTextPainter({required VoidCallback onChanged})
@@ -25,7 +24,7 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
   final TextStyle textStyle;
   final TextAlign textAlign;
 
-  NeumorphicPainterCache _cache;
+  final NeumorphicPainterCache _cache;
 
   late Paint _backgroundPaint;
   late Paint _whiteShadowPaint;
@@ -44,19 +43,19 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
   late ui.Paragraph _gradientParagraph;
 
   void generatePainters() {
-    this._backgroundPaint = Paint();
-    this._whiteShadowPaint = Paint();
-    this._whiteShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
-    this._blackShadowPaint = Paint();
-    this._blackShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
-    this._gradientPaint = Paint();
+    _backgroundPaint = Paint();
+    _whiteShadowPaint = Paint();
+    _whiteShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
+    _blackShadowPaint = Paint();
+    _blackShadowMaskPaint = Paint()..blendMode = BlendMode.dstOut;
+    _gradientPaint = Paint();
 
-    this._borderPaint = Paint()
+    _borderPaint = Paint()
       ..strokeCap = StrokeCap.round
       ..strokeJoin = StrokeJoin.bevel
       ..style = PaintingStyle.stroke
       ..strokeWidth = style.border.width ?? 0.0
-      ..color = style.border.color ?? Color(0xFFFFFFFF);
+      ..color = style.border.color ?? const Color(0xFFFFFFFF);
   }
 
   final bool drawGradient;
@@ -82,29 +81,27 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
   void _updateCache(Offset offset, ImageConfiguration configuration) {
     bool invalidateSize = false;
     if (configuration.size != null) {
-      invalidateSize = this
-          ._cache
+      invalidateSize = _cache
           .updateSize(newOffset: offset, newSize: configuration.size!);
     }
 
-    final bool invalidateLightSource = this
-        ._cache
+    final bool invalidateLightSource = _cache
         .updateLightSource(style.lightSource, style.oppositeShadowLightSource);
 
     bool invalidateColor = false;
     if (style.color != null) {
-      invalidateColor = this._cache.updateStyleColor(style.color!);
+      invalidateColor = _cache.updateStyleColor(style.color!);
       if (invalidateColor) {
-        _backgroundPaint..color = _cache.backgroundColor;
+        _backgroundPaint.color = _cache.backgroundColor;
       }
     }
 
     bool invalidateDepth = false;
     if (style.depth != null) {
-      invalidateDepth = this._cache.updateStyleDepth(style.depth!, 3);
+      invalidateDepth = _cache.updateStyleDepth(style.depth!, 3);
       if (invalidateDepth) {
-        _blackShadowPaint..maskFilter = _cache.maskFilterBlur;
-        _whiteShadowPaint..maskFilter = _cache.maskFilterBlur;
+        _blackShadowPaint.maskFilter = _cache.maskFilterBlur;
+        _whiteShadowPaint.maskFilter = _cache.maskFilterBlur;
       }
     }
 
@@ -112,24 +109,24 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
     if (style.shadowLightColor != null &&
         style.shadowDarkColor != null &&
         style.intensity != null) {
-      invalidateShadowColors = this._cache.updateShadowColor(
+      invalidateShadowColors = _cache.updateShadowColor(
             newShadowLightColorEmboss: style.shadowLightColor!,
             newShadowDarkColorEmboss: style.shadowDarkColor!,
             newIntensity: style.intensity ?? neumorphicDefaultTheme.intensity,
           );
       if (invalidateShadowColors) {
         if (_cache.shadowLightColor != null) {
-          _whiteShadowPaint..color = _cache.shadowLightColor!;
+          _whiteShadowPaint.color = _cache.shadowLightColor!;
         }
         if (_cache.shadowDarkColor != null) {
-          _blackShadowPaint..color = _cache.shadowDarkColor!;
+          _blackShadowPaint.color = _cache.shadowDarkColor!;
         }
       }
     }
 
     final constraints = ui.ParagraphConstraints(width: _cache.width);
     final paragraphStyle = textStyle.getParagraphStyle(
-        textDirection: TextDirection.ltr, textAlign: this.textAlign);
+        textDirection: TextDirection.ltr, textAlign: textAlign);
 
     final textParagraphBuilder = ui.ParagraphBuilder(paragraphStyle)
       ..pushStyle(ui.TextStyle(
@@ -189,8 +186,8 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
             gradientRect: Rect.fromLTRB(0, 0, _cache.width, _cache.height),
             intensity: style.surfaceIntensity,
             source: style.shape == NeumorphicShape.concave
-                ? this.style.lightSource
-                : this.style.lightSource.invert(),
+                ? style.lightSource
+                : style.lightSource.invert(),
           ),
       ))
       ..addText(text);
@@ -222,7 +219,7 @@ class NeumorphicDecorationTextPainter extends BoxPainter {
     if (true) {
       _drawBackground(offset: offset, canvas: canvas, path: path);
     }
-    if (this.drawGradient) {
+    if (drawGradient) {
       _drawGradient(offset: offset, canvas: canvas, path: path);
     }
     if (style.border.isEnabled) {
