@@ -1,11 +1,10 @@
-import '../../flutter_neumorphic.dart';
-import 'neumorphic_path_provider.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 import 'dart:math' as math;
 
 class BeveledPathProvider extends NeumorphicPathProvider {
   final BorderRadius borderRadius;
 
-  const BeveledPathProvider(this.borderRadius, {Listenable? reclip});
+  const BeveledPathProvider(this.borderRadius, {super.reclip});
 
   @override
   bool shouldReclip(NeumorphicPathProvider oldClipper) {
