@@ -1,8 +1,7 @@
-import '../../flutter_neumorphic.dart';
-import 'neumorphic_path_provider.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class RectPathProvider extends NeumorphicPathProvider {
-  const RectPathProvider({Listenable? reclip});
+  const RectPathProvider({super.reclip});
 
   @override
   bool shouldReclip(NeumorphicPathProvider oldClipper) {
