@@ -1,15 +1,7 @@
 import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart' show IconThemeData, TextTheme;
-import 'package:flutter/painting.dart';
-import 'package:flutter_neumorphic/src/theme/app_bar.dart';
-import 'package:flutter_neumorphic/src/widget/container.dart';
 
-import '../../flutter_neumorphic.dart';
-import '../colors.dart';
-import '../light_source.dart';
-import '../shape.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 export '../colors.dart';
 export '../light_source.dart';
@@ -28,13 +20,13 @@ const double _defaultBorderSize = 0.3;
 
 /// Used with the NeumorphicTheme
 ///
-/// ```
+/// ```dart
 /// NeumorphicTheme(
 ///   theme: NeumorphicThemeData(...)
 ///   darkTheme: : NeumorphicThemeData(...)
 ///   child: ...
 /// )`
-/// ``
+/// ```
 ///
 /// Contains all default values used in child Neumorphic Elements as
 /// default colors : baseColor, accentColor, variantColor
@@ -104,9 +96,9 @@ class NeumorphicThemeData {
     this.borderColor = NeumorphicColors.defaultBorder,
     this.borderWidth = _defaultBorderSize,
     this.disableDepth = false,
-  })  : this._depth = depth,
-        this._boxShape = boxShape,
-        this._intensity = intensity;
+  })  : _depth = depth,
+        _boxShape = boxShape,
+        _intensity = intensity;
 
   const NeumorphicThemeData.dark({
     this.baseColor = NeumorphicColors.darkBackground,
@@ -129,9 +121,9 @@ class NeumorphicThemeData {
     this.borderColor = NeumorphicColors.darkDefaultBorder,
     this.borderWidth = _defaultBorderSize,
     this.disableDepth = false,
-  })  : this._depth = depth,
-        this._boxShape = boxShape,
-        this._intensity = intensity;
+  })  : _depth = depth,
+        _boxShape = boxShape,
+        _intensity = intensity;
 
   @override
   String toString() {
@@ -212,7 +204,7 @@ class NeumorphicThemeData {
     double? borderSize,
     LightSource? lightSource,
   }) {
-    return new NeumorphicThemeData(
+    return NeumorphicThemeData(
       baseColor: baseColor ?? this.baseColor,
       textTheme: textTheme ?? this.textTheme,
       iconTheme: iconTheme ?? this.iconTheme,
@@ -230,10 +222,10 @@ class NeumorphicThemeData {
           shadowDarkColorEmboss ?? this.shadowDarkColorEmboss,
       shadowLightColorEmboss:
           shadowLightColorEmboss ?? this.shadowLightColorEmboss,
-      depth: depth ?? this._depth,
-      borderWidth: borderSize ?? this.borderWidth,
+      depth: depth ?? _depth,
+      borderWidth: borderSize ?? borderWidth,
       borderColor: borderColor ?? this.borderColor,
-      intensity: intensity ?? this._intensity,
+      intensity: intensity ?? _intensity,
       lightSource: lightSource ?? this.lightSource,
     );
   }
@@ -243,7 +235,7 @@ class NeumorphicThemeData {
   NeumorphicThemeData copyFrom({
     required NeumorphicThemeData other,
   }) {
-    return new NeumorphicThemeData(
+    return NeumorphicThemeData(
       baseColor: other.baseColor,
       accentColor: other.accentColor,
       variantColor: other.variantColor,
@@ -288,9 +280,9 @@ class NeumorphicBorder {
   });
 
   const NeumorphicBorder.none()
-      : this.isEnabled = true,
-        this.color = const Color(0x00000000),
-        this.width = 0;
+      : isEnabled = true,
+        color = const Color(0x00000000),
+        width = 0;
 
   @override
   bool operator ==(Object other) =>
@@ -325,7 +317,7 @@ class NeumorphicBorder {
 
   NeumorphicBorder copyWithThemeIfNull({Color? color, double? width}) {
     return NeumorphicBorder(
-      isEnabled: this.isEnabled,
+      isEnabled: isEnabled,
       color: this.color ?? color,
       width: this.width ?? width,
     );
@@ -375,10 +367,10 @@ class NeumorphicStyle {
     double surfaceIntensity = 0.25,
     this.disableDepth,
     this.oppositeShadowLightSource = false,
-  })  : this._depth = depth,
-        this.theme = null,
-        this._intensity = intensity,
-        this._surfaceIntensity = surfaceIntensity;
+  })  : _depth = depth,
+        theme = null,
+        _intensity = intensity,
+        _surfaceIntensity = surfaceIntensity;
 
   // with theme constructor is only available privately, please use copyWithThemeIfNull
   const NeumorphicStyle._withTheme({
@@ -397,9 +389,9 @@ class NeumorphicStyle {
     double? depth,
     double? intensity,
     double surfaceIntensity = 0.25,
-  })  : this._depth = depth,
-        this._intensity = intensity,
-        this._surfaceIntensity = surfaceIntensity;
+  })  : _depth = depth,
+        _intensity = intensity,
+        _surfaceIntensity = surfaceIntensity;
 
   double? get depth =>
       _depth?.clamp(Neumorphic.MIN_DEPTH, Neumorphic.MAX_DEPTH);
@@ -413,23 +405,23 @@ class NeumorphicStyle {
   NeumorphicStyle copyWithThemeIfNull(NeumorphicThemeData theme) {
     return NeumorphicStyle._withTheme(
         theme: theme,
-        color: this.color ?? theme.baseColor,
-        boxShape: this.boxShape ?? theme.boxShape,
-        shape: this.shape,
-        border: this.border.copyWithThemeIfNull(
+        color: color ?? theme.baseColor,
+        boxShape: boxShape ?? theme.boxShape,
+        shape: shape,
+        border: border.copyWithThemeIfNull(
             color: theme.borderColor, width: theme.borderWidth),
-        shadowDarkColor: this.shadowDarkColor ?? theme.shadowDarkColor,
-        shadowLightColor: this.shadowLightColor ?? theme.shadowLightColor,
+        shadowDarkColor: shadowDarkColor ?? theme.shadowDarkColor,
+        shadowLightColor: shadowLightColor ?? theme.shadowLightColor,
         shadowDarkColorEmboss:
-            this.shadowDarkColorEmboss ?? theme.shadowDarkColorEmboss,
+            shadowDarkColorEmboss ?? theme.shadowDarkColorEmboss,
         shadowLightColorEmboss:
-            this.shadowLightColorEmboss ?? theme.shadowLightColorEmboss,
-        depth: this.depth ?? theme.depth,
-        intensity: this.intensity ?? theme.intensity,
-        disableDepth: this.disableDepth ?? theme.disableDepth,
-        surfaceIntensity: this.surfaceIntensity,
-        oppositeShadowLightSource: this.oppositeShadowLightSource,
-        lightSource: this.lightSource);
+            shadowLightColorEmboss ?? theme.shadowLightColorEmboss,
+        depth: depth ?? theme.depth,
+        intensity: intensity ?? theme.intensity,
+        disableDepth: disableDepth ?? theme.disableDepth,
+        surfaceIntensity: surfaceIntensity,
+        oppositeShadowLightSource: oppositeShadowLightSource,
+        lightSource: lightSource);
   }
 
   @override
@@ -499,7 +491,7 @@ class NeumorphicStyle {
       shadowLightColorEmboss:
           shadowLightColorEmboss ?? this.shadowLightColorEmboss,
       depth: depth ?? this.depth,
-      theme: this.theme,
+      theme: theme,
       intensity: intensity ?? this.intensity,
       surfaceIntensity: surfaceIntensity ?? this.surfaceIntensity,
       disableDepth: disableDepth ?? this.disableDepth,
@@ -517,7 +509,7 @@ class NeumorphicStyle {
 
   NeumorphicStyle applyDisableDepth() {
     if (disableDepth == true) {
-      return this.copyWith(depth: 0);
+      return copyWith(depth: 0);
     } else {
       return this;
     }
