@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 class NeumorphicBackButton extends StatelessWidget {
@@ -9,12 +7,12 @@ class NeumorphicBackButton extends StatelessWidget {
   final bool forward;
 
   const NeumorphicBackButton({
-    Key? key,
+    super.key,
     this.onPressed,
     this.style,
     this.padding,
     this.forward = false,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
