@@ -7,7 +7,7 @@ import 'package:flutter_neumorphic/src/theme/neumorphic_theme.dart';
 ///
 /// It can provide too a roundRect clip of the screen border using [borderRadius], [margin] and [backendColor]
 ///
-/// ```
+/// ```dart
 /// NeumorphicBackground(
 ///      borderRadius: BorderRadius.circular(12),
 ///      margin: EdgeInsets.all(12),
@@ -33,15 +33,15 @@ class NeumorphicBackground extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: this.margin,
-      color: this.backendColor,
+      padding: margin,
+      color: backendColor,
       child: ClipRRect(
-        borderRadius: this.borderRadius ?? BorderRadius.circular(0),
+        borderRadius: borderRadius ?? BorderRadius.circular(0),
         child: AnimatedContainer(
           color: NeumorphicTheme.baseColor(context),
-          padding: this.padding,
+          padding: padding,
           duration: const Duration(milliseconds: 100),
-          child: this.child,
+          child: child,
         ),
       ),
     );
