@@ -1,13 +1,8 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_neumorphic/flutter_neumorphic.dart';
-import '../theme/neumorphic_theme.dart';
-import '../widget/app_bar.dart';
-import 'animation/animated_scale.dart' as animationScale;
-import 'container.dart';
+import 'package:flutter_neumorphic/src/widget/animation/animated_scale.dart' as animation_scale;
 
-typedef void NeumorphicButtonClickListener();
+typedef NeumorphicButtonClickListener = void Function();
 
 /// A Neumorphic Button
 ///
@@ -26,7 +21,7 @@ typedef void NeumorphicButtonClickListener();
 ///
 /// It takes a [NeumorphicStyle] @see [Neumorphi]
 ///
-/// ```
+/// ```dart
 ///  NeumorphicButton(
 ///          onClick: () {
 ///            setState(() {
@@ -59,8 +54,8 @@ class NeumorphicButton extends StatefulWidget {
   final bool provideHapticFeedback;
   final String? tooltip;
 
-  NeumorphicButton({
-    Key? key,
+  const NeumorphicButton({
+    super.key,
     this.padding,
     this.margin = EdgeInsets.zero,
     this.child,
@@ -74,7 +69,7 @@ class NeumorphicButton extends StatefulWidget {
     this.minDistance = 0,
     this.style,
     this.provideHapticFeedback = true,
-  }) : super(key: key);
+  });
 
   bool get isEnabled => onPressed != null;
 
@@ -92,7 +87,7 @@ class _NeumorphicButtonState extends State<NeumorphicButton> {
     final appBarPresent = NeumorphicAppBarTheme.of(context) != null;
 
     final theme = NeumorphicTheme.currentTheme(context);
-    this.initialStyle = widget.style ??
+    initialStyle = widget.style ??
         (appBarPresent
             ? theme.appBarTheme.buttonStyle
             : (theme.buttonStyle ?? const NeumorphicStyle()));
@@ -166,7 +161,7 @@ class _NeumorphicButtonState extends State<NeumorphicButton> {
     final result = _build(context);
     if (widget.tooltip != null) {
       return Tooltip(
-        message: widget.tooltip!,
+        message: widget.tooltip,
         child: result,
       );
     } else {
@@ -196,10 +191,10 @@ class _NeumorphicButtonState extends State<NeumorphicButton> {
         hasTapUp = true;
         _resetIfTapUp();
       },
-      child: animationScale.AnimatedScale(
+      child: animation_scale.AnimatedScale(
         scale: _getScale(),
         child: Neumorphic(
-          margin: widget.margin ?? const EdgeInsets.all(0),
+          margin: widget.margin ?? EdgeInsets.zero,
           drawSurfaceAboveChild: widget.drawSurfaceAboveChild,
           duration: widget.duration,
           curve: widget.curve,
@@ -217,7 +212,7 @@ class _NeumorphicButtonState extends State<NeumorphicButton> {
 
   double _getDepth() {
     if (widget.isEnabled) {
-      return this.depth;
+      return depth;
     } else {
       return 0;
     }
@@ -230,7 +225,7 @@ class _NeumorphicButtonState extends State<NeumorphicButton> {
           ? NeumorphicButton.PRESSED_SCALE
           : NeumorphicButton.UNPRESSED_SCALE;
     } else {
-      return this.pressed
+      return pressed
           ? NeumorphicButton.PRESSED_SCALE
           : NeumorphicButton.UNPRESSED_SCALE;
     }
