@@ -1,12 +1,12 @@
 import 'package:flutter/widgets.dart';
 
-import '../../neumorphic_box_shape.dart';
+import 'package:flutter_neumorphic/src/neumorphic_box_shape.dart';
 
 class NeumorphicBoxShapeClipper extends StatelessWidget {
   final NeumorphicBoxShape shape;
   final Widget? child;
 
-  NeumorphicBoxShapeClipper({required this.shape, this.child});
+  const NeumorphicBoxShapeClipper({required this.shape, this.child});
 
   CustomClipper<Path>? _getClipper(NeumorphicBoxShape shape) {
     return shape.customShapePathProvider;
@@ -15,7 +15,7 @@ class NeumorphicBoxShapeClipper extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ClipPath(
-      clipper: _getClipper(this.shape),
+      clipper: _getClipper(shape),
       child: child,
     );
   }
