@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart' as material;
 import 'package:flutter/widgets.dart';
 
-import '../neumorphic_box_shape.dart';
-import '../decoration/neumorphic_decorations.dart';
-import '../theme/neumorphic_theme.dart';
-import 'clipper/neumorphic_box_shape_clipper.dart';
+import 'package:flutter_neumorphic/src/neumorphic_box_shape.dart';
+import 'package:flutter_neumorphic/src/decoration/neumorphic_decorations.dart';
+import 'package:flutter_neumorphic/src/theme/neumorphic_theme.dart';
+import 'package:flutter_neumorphic/src/widget/clipper/neumorphic_box_shape_clipper.dart';
 
 export '../neumorphic_box_shape.dart';
 export '../decoration/neumorphic_decorations.dart';
@@ -36,8 +36,8 @@ export '../theme/neumorphic_theme.dart';
 ///
 @immutable
 class Neumorphic extends StatelessWidget {
-  static const DEFAULT_DURATION = const Duration(milliseconds: 100);
-  static const DEFAULT_CURVE = Curves.linear;
+  static const Duration DEFAULT_DURATION = Duration(milliseconds: 100);
+  static const material.Curve DEFAULT_CURVE = Curves.linear;
 
   static const double MIN_DEPTH = -20.0;
   static const double MAX_DEPTH = 20.0;
@@ -59,34 +59,34 @@ class Neumorphic extends StatelessWidget {
   final bool
       drawSurfaceAboveChild; //if true => boxDecoration & foreground decoration, else => boxDecoration does all the work
 
-  Neumorphic({
-    Key? key,
+  const Neumorphic({
+    super.key,
     this.child,
     this.duration = Neumorphic.DEFAULT_DURATION,
     this.curve = Neumorphic.DEFAULT_CURVE,
     this.style,
     this.textStyle,
-    this.margin = const EdgeInsets.all(0),
-    this.padding = const EdgeInsets.all(0),
+    this.margin = EdgeInsets.zero,
+    this.padding = EdgeInsets.zero,
     this.drawSurfaceAboveChild = true,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     final theme = NeumorphicTheme.currentTheme(context);
-    final NeumorphicStyle style = (this.style ?? NeumorphicStyle())
+    final NeumorphicStyle style = (this.style ?? const NeumorphicStyle())
         .copyWithThemeIfNull(theme)
         .applyDisableDepth();
 
     return _NeumorphicContainer(
-      padding: this.padding,
-      textStyle: this.textStyle,
-      drawSurfaceAboveChild: this.drawSurfaceAboveChild,
-      duration: this.duration,
+      padding: padding,
+      textStyle: textStyle,
+      drawSurfaceAboveChild: drawSurfaceAboveChild,
+      duration: duration,
       style: style,
-      curve: this.curve,
-      margin: this.margin,
-      child: this.child,
+      curve: curve,
+      margin: margin,
+      child: child,
     );
   }
 }
@@ -101,8 +101,9 @@ class _NeumorphicContainer extends StatelessWidget {
   final bool drawSurfaceAboveChild;
   final EdgeInsets padding;
 
-  _NeumorphicContainer({
-    Key? key,
+  const _NeumorphicContainer({
+    // ignore: unused_element_parameter
+    super.key,
     this.child,
     this.textStyle,
     required this.padding,
@@ -111,37 +112,37 @@ class _NeumorphicContainer extends StatelessWidget {
     required this.curve,
     required this.style,
     required this.drawSurfaceAboveChild,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
-    final shape = this.style.boxShape ?? NeumorphicBoxShape.rect();
+    final shape = style.boxShape ?? const NeumorphicBoxShape.rect();
 
     return DefaultTextStyle(
-      style: this.textStyle ?? material.Theme.of(context).textTheme.bodyMedium!,
+      style: textStyle ?? material.Theme.of(context).textTheme.bodyMedium!,
       child: AnimatedContainer(
-        margin: this.margin,
-        duration: this.duration,
-        curve: this.curve,
+        margin: margin,
+        duration: duration,
+        curve: curve,
         child: NeumorphicBoxShapeClipper(
           shape: shape,
           child: Padding(
-            padding: this.padding,
-            child: this.child,
+            padding: padding,
+            child: child,
           ),
         ),
         foregroundDecoration: NeumorphicDecoration(
           isForeground: true,
           renderingByPath: shape.customShapePathProvider.oneGradientPerPath,
-          splitBackgroundForeground: this.drawSurfaceAboveChild,
-          style: this.style,
+          splitBackgroundForeground: drawSurfaceAboveChild,
+          style: style,
           shape: shape,
         ),
         decoration: NeumorphicDecoration(
           isForeground: false,
           renderingByPath: shape.customShapePathProvider.oneGradientPerPath,
-          splitBackgroundForeground: this.drawSurfaceAboveChild,
-          style: this.style,
+          splitBackgroundForeground: drawSurfaceAboveChild,
+          style: style,
           shape: shape,
         ),
       ),
