@@ -1,4 +1,4 @@
-import '../../flutter_neumorphic.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 const BoxConstraints _kSizeConstraints = BoxConstraints.tightFor(
   width: 56.0,
@@ -18,25 +18,25 @@ class NeumorphicFloatingActionButton extends StatelessWidget {
   final NeumorphicStyle? style;
 
   const NeumorphicFloatingActionButton({
-    Key? key,
+    super.key,
     this.mini = false,
     this.style,
     this.tooltip,
     @required this.child,
     @required this.onPressed,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     return ConstrainedBox(
-      constraints: this.mini ? _kMiniSizeConstraints : _kSizeConstraints,
+      constraints: mini ? _kMiniSizeConstraints : _kSizeConstraints,
       child: NeumorphicButton(
-        padding: EdgeInsets.all(0),
-        onPressed: this.onPressed,
-        tooltip: this.tooltip,
-        style: this.style ??
+        padding: EdgeInsets.zero,
+        onPressed: onPressed,
+        tooltip: tooltip,
+        style: style ??
             NeumorphicTheme.currentTheme(context).appBarTheme.buttonStyle,
-        child: this.child,
+        child: child,
       ),
     );
   }
