@@ -1,7 +1,5 @@
-import 'package:flutter/widgets.dart';
 
-import '../../flutter_neumorphic.dart';
-import '../theme/neumorphic_theme.dart';
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 export '../decoration/neumorphic_decorations.dart';
 export '../neumorphic_box_shape.dart';
@@ -15,14 +13,14 @@ class NeumorphicIcon extends StatelessWidget {
   final double size;
   final Duration duration;
 
-  NeumorphicIcon(
+  const NeumorphicIcon(
     this.icon, {
-    Key? key,
+    super.key,
     this.duration = Neumorphic.DEFAULT_DURATION,
     this.curve = Neumorphic.DEFAULT_CURVE,
     this.style,
     this.size = 20,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -33,9 +31,9 @@ class NeumorphicIcon extends StatelessWidget {
         fontFamily: icon.fontFamily,
         package: icon.fontPackage,
       ),
-      duration: this.duration,
+      duration: duration,
       style: style,
-      curve: this.curve,
+      curve: curve,
     );
   }
 }
