@@ -1,11 +1,8 @@
-import 'package:flutter/widgets.dart';
-import 'package:flutter_neumorphic/src/widget/animation/animated_scale.dart'
-    as animationScale;
 
-import '../../flutter_neumorphic.dart';
-import '../neumorphic_box_shape.dart';
-import '../theme/neumorphic_theme.dart';
-import 'container.dart';
+import 'package:flutter_neumorphic/src/widget/animation/animated_scale.dart'
+    as animation_scale;
+
+import 'package:flutter_neumorphic/flutter_neumorphic.dart';
 
 /// A style to customize the [NeumorphicSwitch]
 ///
@@ -79,7 +76,7 @@ class NeumorphicSwitchStyle {
 /// The switch itself does not maintain any state. Instead, when the state of the switch changes, the widget calls the onChanged callback.
 /// Most widgets that use a switch will listen for the onChanged callback and rebuild the switch with a new value to update the visual appearance of the switch.
 ///
-/// ```
+/// ```dart
 ///  bool _switch1Value = false;
 ///  bool _switch2Value = false;
 ///
@@ -128,54 +125,52 @@ class NeumorphicSwitch extends StatelessWidget {
 
   const NeumorphicSwitch({
     this.style = const NeumorphicSwitchStyle(),
-    Key? key,
+    super.key,
     this.curve = Neumorphic.DEFAULT_CURVE,
     this.duration = const Duration(milliseconds: 200),
     this.value = false,
     this.onChanged,
     this.height = 40,
     this.isEnabled = true,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     final NeumorphicThemeData theme = NeumorphicTheme.currentTheme(context);
     return SizedBox(
-      height: this.height,
+      height: height,
       child: AspectRatio(
         aspectRatio: 2 / 1,
         child: GestureDetector(
           onTap: () {
             // animation breaking prevention
-            if (!this.isEnabled) {
+            if (!isEnabled) {
               return;
             }
-            if (this.onChanged != null) {
-              this.onChanged!(!this.value);
-            }
+            onChanged?.call(!value);
           },
           child: Neumorphic(
             drawSurfaceAboveChild: false,
             style: NeumorphicStyle(
-              boxShape: NeumorphicBoxShape.stadium(),
-              lightSource: this.style.lightSource ?? theme.lightSource,
-              border: this.style.trackBorder,
-              disableDepth: this.style.disableDepth,
+              boxShape: const NeumorphicBoxShape.stadium(),
+              lightSource: style.lightSource ?? theme.lightSource,
+              border: style.trackBorder,
+              disableDepth: style.disableDepth,
               depth: _getTrackDepth(theme.depth),
-              shape: NeumorphicShape.flat,
-              color: _getTrackColor(theme, this.isEnabled),
+              //shape: NeumorphicShape.flat,
+              color: _getTrackColor(theme, isEnabled),
             ),
-            child: animationScale.AnimatedScale(
-              scale: this.isEnabled ? 1 : 0,
-              alignment: this.value ? Alignment(0.5, 0) : Alignment(-0.5, 0),
+            child: animation_scale.AnimatedScale(
+              scale: isEnabled ? 1 : 0,
+              alignment: value ? const Alignment(0.5, 0) : const Alignment(-0.5, 0),
               child: AnimatedThumb(
-                curve: this.curve,
-                disableDepth: this.style.disableDepth,
-                depth: this._thumbDepth,
-                duration: this.duration,
-                alignment: this._alignment,
+                curve: curve,
+                disableDepth: style.disableDepth,
+                depth: _thumbDepth,
+                duration: duration,
+                alignment: _alignment,
                 shape: _getThumbShape,
-                lightSource: this.style.lightSource ?? theme.lightSource,
+                lightSource: style.lightSource ?? theme.lightSource,
                 border: style.thumbBorder,
                 thumbColor: _getThumbColor(theme),
               ),
@@ -187,7 +182,7 @@ class NeumorphicSwitch extends StatelessWidget {
   }
 
   Alignment get _alignment {
-    if (this.value) {
+    if (value) {
       return Alignment.centerRight;
     } else {
       return Alignment.centerLeft;
@@ -195,37 +190,38 @@ class NeumorphicSwitch extends StatelessWidget {
   }
 
   double get _thumbDepth {
-    if (!this.isEnabled) {
+    if (!isEnabled) {
       return 0;
-    } else
-      return this.style.thumbDepth ?? neumorphicDefaultTheme.depth;
+    } else {
+      return style.thumbDepth ?? neumorphicDefaultTheme.depth;
+    }
   }
 
   NeumorphicShape get _getThumbShape {
-    return this.style.thumbShape ?? NeumorphicShape.flat;
+    return style.thumbShape ?? NeumorphicShape.flat;
   }
 
   double? _getTrackDepth(double? themeDepth) {
     if (themeDepth == null) return themeDepth;
     //force negative to have emboss
-    final double depth = -1 * (this.style.trackDepth ?? themeDepth).abs();
+    final double depth = -1 * (style.trackDepth ?? themeDepth).abs();
     return depth.clamp(Neumorphic.MIN_DEPTH, NeumorphicSwitch.MIN_EMBOSS_DEPTH);
   }
 
   Color _getTrackColor(NeumorphicThemeData theme, bool enabled) {
     if (!enabled) {
-      return this.style.inactiveTrackColor ?? theme.baseColor;
+      return style.inactiveTrackColor ?? theme.baseColor;
     }
 
-    return this.value == true
-        ? this.style.activeTrackColor ?? theme.accentColor
-        : this.style.inactiveTrackColor ?? theme.baseColor;
+    return value == true
+        ? style.activeTrackColor ?? theme.accentColor
+        : style.inactiveTrackColor ?? theme.baseColor;
   }
 
   Color _getThumbColor(NeumorphicThemeData theme) {
-    Color? color = this.value == true
-        ? this.style.activeThumbColor
-        : this.style.inactiveThumbColor;
+    final Color? color = value == true
+        ? style.activeThumbColor
+        : style.inactiveThumbColor;
     return color ?? theme.baseColor;
   }
 }
@@ -241,8 +237,8 @@ class AnimatedThumb extends StatelessWidget {
   final NeumorphicBorder border;
   final LightSource lightSource;
 
-  AnimatedThumb({
-    Key? key,
+  const AnimatedThumb({
+    super.key,
     this.thumbColor,
     required this.alignment,
     required this.duration,
@@ -252,26 +248,26 @@ class AnimatedThumb extends StatelessWidget {
     this.border = const NeumorphicBorder.none(),
     this.lightSource = LightSource.topLeft,
     this.disableDepth = false,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     // This Container is actually the inner track containing the thumb
     return AnimatedAlign(
-      curve: this.curve,
-      alignment: this.alignment,
-      duration: this.duration,
+      curve: curve,
+      alignment: alignment,
+      duration: duration,
       child: Padding(
         padding: const EdgeInsets.all(4.0),
         child: Neumorphic(
           style: NeumorphicStyle(
-            boxShape: NeumorphicBoxShape.circle(),
-            disableDepth: this.disableDepth,
+            boxShape: const NeumorphicBoxShape.circle(),
+            disableDepth: disableDepth,
             shape: shape,
-            depth: this.depth,
+            depth: depth,
             color: thumbColor,
-            border: this.border,
-            lightSource: this.lightSource,
+            border: border,
+            lightSource: lightSource,
           ),
           child: AspectRatio(
             aspectRatio: 1,
